Scripts/requirements.txt
yahooquery>=2.4.1
pandas>=2.3.3
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0